% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean_strings}
\alias{clean_strings}
\title{Function to trim then convert empty strings in data.frame into NAs.}
\usage{
clean_strings(input_dataframe)
}
\arguments{
\item{input_dataframe}{A data.frame object}
}
\description{
This function takes a data.frame object, trims all character strings, coverts empty strings like '' and "" into NA, then outputs the cleaned data.frame.
}
\examples{
clean_strings()
}
