% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_all_na_col.R
\name{drop_all_na_col}
\alias{drop_all_na_col}
\title{Function to drop columns that have only NAs.}
\usage{
drop_all_na_col(input_dataframe)
}
\arguments{
\item{input_dataframe}{A data.frame object}
}
\description{
This function takes a data.frame object and outputs the data.frame with only columns that contain at least 1 non-NA value. Best when used with clean_strings() to make sure empty cells are appropriately identified.
}
\examples{
drop_all_na_col()
}
