% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_column_labels.R
\name{extract_column_labels}
\alias{extract_column_labels}
\title{Extract labels for each column in a data frame}
\usage{
extract_column_labels(input_df, which_string = "label")
}
\arguments{
\item{input_df}{A data.frame() object.}

\item{which_string}{String fed into the `which` argument of the attr() function. Defaults to "label".}
}
\description{
This function returns a list of the labels for all columns in a dataframe.
}
