% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pvalue.R
\name{format_pvalue}
\alias{format_pvalue}
\title{Function to make p-values prettier for reports.}
\usage{
format_pvalue(pvalue, out_format = "string")
}
\arguments{
\item{pvalue}{A single, numeric p-value.}

\item{out_format}{Specifies what will be returned. Defaults to strings. Indicating numeric will return 0 for values less than 0.0001.}
}
\description{
This function rounds individual p-values to 4 digits and returns them to the user.
}
\examples{
> p_value = 0.000000345
> format_pvalue(p_value)
[1] "<0.0001"

################

> p_value = 0.000000345
> format_pvalue(p_value,"numeric")
[1] 0
}
