% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libraries.R
\name{load_libraries}
\alias{load_libraries}
\title{A library loading function.}
\usage{
load_libraries(libraries = c("tidyverse", "magrittr", "haven", "ggplot2"))
}
\arguments{
\item{libraries}{A vector of library names as strings. Defaults to "tidyverse","magrittr","haven", and "ggplot2"}
}
\description{
This function loads a list of libraries. If any of the libraries are missing, it installs them first.
}
\examples{
load_libraries()
}
\keyword{library}
