% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_codes_with_labels.R
\name{replace_codes_with_labels}
\alias{replace_codes_with_labels}
\title{Replace codes in a haven_labelled vector to the true values.}
\usage{
replace_codes_with_labels(labelled_vector, label_string = "labels")
}
\arguments{
\item{labelled_vector}{An object of class haven_labelled. Can confirm using: `is(vector,"haven_labelled")`}

\item{label_string}{String fed into the `which` argument of the attr() function. Defaults to "labels".}
}
\description{
This function replaces the codes in a haven_labelled vector with the values associated with the code. Can be used in a mutate call from dplyr or on any appropriate vector.
}
\examples{
haven_loaded_data_frame \%>\%
  mutate(new_column = replace_codes_with_labels(haven_labelled_vector))
}
