% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_HPRC.R
\name{update_HPRC}
\alias{update_HPRC}
\title{Update the HPRC file from github}
\usage{
update_HPRC()
}
\description{
This function downloads the most recent HPRC package file from GitHub.
}
\examples{
update_HPRC()
}
